requirements.txt
# Connecteurs & API
python-dotenv
requests
requests-mock
geopy

# Data Science
numpy==2.2.0
pandas
scikit-learn
xgboost
prophet
shap
faiss-cpu
sentence-transformers
transformers
sentencepiece
tqdm

# Visualisation
matplotlib
seaborn
plotly

# Jupyter Notebooks
jupyter
ipykernel
nbformat>=4.2.0

# Backend API
fastapi
uvicorn
flask
gunicorn

# Frontend & autres
streamlit
joblib
openpyxl
statsmodels
pmdarima==2.0.3

# Tests
pytest

# Qualité du code
black
isort
flake8
pre-commit
